chore: refactor assets structure
